Refactor asgard_resources to remove circular dependency
